+++ sportorg/modules/teamwork/packet_header.py
@@ -81,6 +81,8 @@ class Header:
 
     header_struck = "=2s2H36sLQ"
     header_size = struct.calcsize(header_struck)
+    VERSION_PLAIN = 0
+    VERSION_AES256_GCM = 1
 
     def __init__(self, obj_data=None, op_type=Operations.Update.name):
         self.pack_tag = b"SO"
@@ -137,8 +139,10 @@ class Header:
         self.size = len(obj_data)
         return True
 
-    def pack_header(self, psize):
+    def pack_header(self, psize, version=None):
         self.size = psize
+        if version is not None:
+            self.version = version
         return struct.pack(
             Header.header_struck,
             self.pack_tag,